RAG API/requirements.txt
fastapi
uvicorn
IPython
boto3
Mangum
PyYAML
langchain
langchain-core
langchain-community
langchainhub
langchain-chroma
langchain-openai
langchain-experimental
langchain-text-splitters
langsmith
boto3
requests
requests-aws4auth
opensearch-py
elasticsearch
beautifulsoup4
openai
python-multipart
pillow
azure-storage-blob
azure-storage-blob
redis
langchain_redis
PyPDF2
rank-bm25
PyMuPDF
python-docx
openpyxl
